requirement.txt
rouge
rouge_score
transformers==4.21.0
pytorch_lightning==1.6.0
pandas
numpy
pytorch==1.11.0